tidy icons in separate folder; fix GernerCard style mistake
