+++ app/components/Button/stories/Icon.stories.mdx
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 import { IconButton } from '..';
-import * as Icons from '../../../staticData/images/icons';
+import * as Icons from '../../../staticData/Svgs/IconButton_Icons';
 
 <Meta title="Components|Button/icon" component={IconButton} />
