+++ app/staticData/Svgs/IconButton_Icons.js
@@ -92,21 +92,6 @@ export const clapIcon = (
   </svg>
 );
 
-export const addMediaPlusIcon = (
-  <svg width="25" height="25">
-    <path d="M20 12h-7V5h-1v7H5v1h7v7h1v-7h7" fillRule="evenodd" />
-  </svg>
-);
-
-export const addCheckboxTickIcon = (
-  <svg width="21" height="21">
-    <path
-      d="M5.5 7.854L3.379 9.975l5.82 5.82 8.675-8.675-2.121-2.121-6.554 6.553z"
-      fillRule="evenodd"
-    />
-  </svg>
-);
-
 export const clapSmallIcon = (
   <svg width="25" height="25" viewBox="0 0 25 25" fill="inherit">
     <path
