+++ app/staticData/Svgs/starIcon.js
